Crate qt_generator_common

Common utilities for the generator and the build script for Qt crates.

Qt modules are identified within this crate using snake case names without a prefix, e.g. core for QtCore and ui_tools for QtUiTools. sublib_name argument should be in this form.

See README for more information.

Structs

InstallationData
Properties of a Qt installation

Functions

framework_name
Returns MacOS framework name of the specified module as should be passed to the linker, e.g. "QtCore".

get_installation_data
Detects properties of current Qt installation using qmake command line utility.

lib_dependencies
Returns list of modules this module depends on.

lib_folder_name
Returns name of the module's include directory, e.g. "QtCore".

real_lib_name
Returns library name of the specified module as should be passed to the linker, e.g. "Qt5Core".