Struct qt_generator_common::InstallationData

pub struct InstallationData {
    pub qt_version: String,
    pub root_include_path: PathBuf,
    pub lib_include_path: PathBuf,
    pub lib_path: PathBuf,
    pub docs_path: PathBuf,
    pub is_framework: bool,
}

Properties of a Qt installation

Fields
qt_version: String

Qt version.

root_include_path: PathBuf

Path to the parent include directory of the installation.

lib_include_path: PathBuf

Path to the include directory of the library that is being processed. This is a direct subdirectory of root_include_path.

lib_path: PathBuf

Path to the directory containing library files for the linker.

docs_path: PathBuf

Path to the directory containing Qt documentation files.

is_framework: bool

If true, this Qt library was built as a MacOS framework.